# api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CS2KZ API",
    "description": "Source Code available on [GitHub](https://github.com/KZGlobalTeam/cs2kz-api).",
    "contact": {
      "name": "AlphaKeks",
      "email": "alphakeks@dawn.sh"
    },
    "license": {
      "name": "Licensed under the GPLv3",
      "url": "https://www.gnu.org/licenses/gpl-3.0"
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/jumpstats": {
      "get": {
        "tags": [
          "Jumpstats"
        ],
        "operationId": "get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by jump type.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/JumpType"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "mode",
            "in": "query",
            "description": "Filter by mode.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Mode"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "style",
            "in": "query",
            "description": "Filter by style.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Style"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "minimum_distance",
            "in": "query",
            "description": "Filter by a minimum distance.",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            }
          },
          {
            "name": "player",
            "in": "query",
            "description": "Filter by player.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/PlayerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "server",
            "in": "query",
            "description": "Filter by server.",
            "required": false,
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/ServerIdentifier"
                }
              ],
              "nullable": true
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "description": "Filter by creation date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "description": "Filter by creation date.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limit the number of returned results.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Paginate by `offset` entries.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A jumpstat.",
                  "required": [
                    "id",
                    "type",
                    "mode",
                    "style",
                    "player",
                    "server",
                    "strafes",
                    "distance",
                    "sync",
                    "pre",
                    "max",
                    "overlap",
                    "bad_angles",
                    "dead_air",
                    "height",
                    "airpath",
                    "deviation",
                    "average_width",
                    "airtime",
                    "created_on"
                  ],
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "uint64",
                      "description": "The jumpstat's ID.",
                      "minimum": 0
                    },
                    "type": {
                      "$ref": "#/components/schemas/JumpType"
                    },
                    "mode": {
                      "$ref": "#/components/schemas/Mode"
                    },
                    "style": {
                      "$ref": "#/components/schemas/Style"
                    },
                    "player": {
                      "$ref": "#/components/schemas/Player"
                    },
                    "server": {
                      "$ref": "#/components/schemas/ServerInfo"
                    },
                    "strafes": {
                      "type": "integer",
                      "format": "uint8",
                      "description": "The amount of strafes done in this jump.",
                      "minimum": 0
                    },
                    "distance": {
                      "type": "number",
                      "format": "float",
                      "description": "The jump's distance."
                    },
                    "sync": {
                      "type": "number",
                      "format": "float",
                      "description": "The % of how much airtime was spent gaining speed."
                    },
                    "pre": {
                      "type": "number",
                      "format": "float",
                      "description": "The jump's speed at jumpoff."
                    },
                    "max": {
                      "type": "number",
                      "format": "float",
                      "description": "The maximum speed during the jump."
                    },
                    "overlap": {
                      "type": "number",
                      "format": "float",
                      "description": "The % of how much airtime was spent pressing both directional keys at once."
                    },
                    "bad_angles": {
                      "type": "number",
                      "format": "float",
                      "description": "The % of how much airtime keys were pressed but no speed was gained."
                    },
                    "dead_air": {
                      "type": "number",
                      "format": "float",
                      "description": "The % of how much airtime was spent not gaining speed."
                    },
                    "height": {
                      "type": "number",
                      "format": "float",
                      "description": "The maximum height during this jump (in units)."
                    },
                    "airpath": {
                      "type": "number",
                      "format": "float",
                      "description": "How close to a perfect airpath this jump was.\n\nThe closer to 1.0 the better."
                    },
                    "deviation": {
                      "type": "number",
                      "format": "float",
                      "description": "How far the landing point deviates from the jumpoff point."
                    },
                    "average_width": {
                      "type": "number",
                      "format": "float",
                      "description": "The average strafe width."
                    },
                    "airtime": {
                      "type": "string",
                      "description": "How much time the player spent in the air."
                    },
                    "created_on": {
                      "type": "string",
                      "format": "date-time",
                      "description": "When this jump was submitted."
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Jumpstats"
        ],
        "operationId": "post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewJumpstat"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "A newly created jumpstat.",
                  "required": [
                    "jumpstat_id"
                  ],
                  "properties": {
                    "jumpstat_id": {
                      "type": "integer",
                      "format": "uint64",
                      "description": "The jumpstat's ID.",
                      "minimum": 0
                    }
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": ""
          },
          "401": {
            "description": ""
          },
          "422": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "security": [
          {
            "CS2 Server": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JumpType": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "longjump",
              "single_bhop",
              "multi_bhop",
              "weirdjump",
              "ladderjump",
              "ladderhop"
            ],
            "example": "longjump"
          },
          {
            "type": "integer",
            "title": "ID",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6
            ],
            "example": 1
          }
        ],
        "example": "longjump"
      },
...
      "Limit": {
        "type": "number",
        "description": "limits the amount of returned values",
        "minimum": 0
      },
...
      "Mode": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "vanilla",
              "classic"
            ],
            "example": "classic"
          },
          {
            "type": "integer",
            "title": "ID",
            "enum": [
              1,
              2
            ],
            "example": 1
          }
        ],
        "example": "classic"
      },
...
      "NewJumpstat": {
        "type": "object",
        "description": "Request body for submitting new jumpstats.",
        "required": [
          "type",
          "mode",
          "style",
          "player_id",
          "strafes",
          "distance",
          "sync",
          "pre",
          "max",
          "overlap",
          "bad_angles",
          "dead_air",
          "height",
          "airpath",
          "deviation",
          "average_width",
          "airtime"
        ],
        "properties": {
          "type": {
            "$ref": "#/components/schemas/JumpType"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "style": {
            "$ref": "#/components/schemas/Style"
          },
          "player_id": {
            "$ref": "#/components/schemas/SteamID"
          },
          "strafes": {
            "type": "integer",
            "format": "uint8",
            "description": "The amount of strafes done in this jump.",
            "minimum": 0
          },
          "distance": {
            "type": "number",
            "format": "float",
            "description": "The jump's distance."
          },
          "sync": {
            "type": "number",
            "format": "float",
            "description": "The % of how much airtime was spent gaining speed."
          },
          "pre": {
            "type": "number",
            "format": "float",
            "description": "The jump's speed at jumpoff."
          },
          "max": {
            "type": "number",
            "format": "float",
            "description": "The maximum speed during the jump."
          },
          "overlap": {
            "type": "number",
            "format": "float",
            "description": "The % of how much airtime was spent pressing both directional keys at once."
          },
          "bad_angles": {
            "type": "number",
            "format": "float",
            "description": "The % of how much airtime keys were pressed but no speed was gained."
          },
          "dead_air": {
            "type": "number",
            "format": "float",
            "description": "The % of how much airtime was spent not gaining speed."
          },
          "height": {
            "type": "number",
            "format": "float",
            "description": "The maximum height during this jump (in units)."
          },
          "airpath": {
            "type": "number",
            "format": "float",
            "description": "How close to a perfect airpath this jump was.\n\nThe closer to 1.0 the better."
          },
          "deviation": {
            "type": "number",
            "format": "float",
            "description": "How far the landing point deviates from the jumpoff point."
          },
          "average_width": {
            "type": "number",
            "format": "float",
            "description": "The average strafe width."
          },
          "airtime": {
            "type": "string",
            "description": "How much time the player spent in the air."
          }
        }
      },
...
      "Offset": {
        "type": "number",
        "description": "used for pagination",
        "default": 0,
        "maximum": 9223372036854775807,
        "minimum": -9223372036854775808
      },
      "Player": {
        "type": "object",
        "description": "A KZ player.",
        "required": [
          "name",
          "steam_id"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The player's name."
          },
          "steam_id": {
            "$ref": "#/components/schemas/SteamID"
          }
        }
      },
      "PlayerIdentifier": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/SteamID"
          },
          {
            "type": "string",
            "description": "A player's name."
          }
        ],
        "description": "Players are usually identified by their [SteamID], which is unique, or their name."
      },
...
      "ServerIdentifier": {
        "oneOf": [
          {
            "type": "integer",
            "format": "uint16",
            "description": "A server's ID.",
            "minimum": 0
          },
          {
            "type": "string",
            "description": "A server's name."
          }
        ],
        "description": "Servers are either identified by their ID, which is unique, or their name."
      },
      "ServerInfo": {
        "type": "object",
        "description": "Information about a server.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint16",
            "description": "The server's ID.",
            "minimum": 0
          },
          "name": {
            "type": "string",
            "description": "The server's name."
          }
        }
      },
...
      "SteamID": {
        "anyOf": [
          {
            "type": "string",
            "title": "Steam ID",
            "example": "STEAM_1:1:161178172"
          },
          {
            "type": "string",
            "title": "Steam ID3",
            "example": "U:1:322356345"
          },
          {
            "type": "integer",
            "title": "Steam ID32",
            "example": 322356345
          },
          {
            "type": "integer",
            "title": "Steam ID64",
            "example": 76561198282622073
          }
        ],
        "description": "a player's SteamID",
        "example": "STEAM_1:1:161178172"
      },
      "Style": {
        "anyOf": [
          {
            "type": "string",
            "title": "Name",
            "enum": [
              "normal",
              "backwards",
              "sideways",
              "half_sideways",
              "w_only",
              "low_gravity",
              "high_gravity",
              "no_prestrafe",
              "negev",
              "ice"
            ],
            "example": "normal"
          },
          {
            "type": "integer",
            "title": "ID",
            "enum": [
              1,
              2,
              3,
              4,
              5,
              6,
              7,
              8,
              9,
              10
            ],
            "example": 1
          }
        ],
        "example": "normal"
      },
...
    },
    "securitySchemes": {
      "API Key": {
        "type": "http",
        "scheme": "bearer"
      },
      "Browser Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "kz-auth"
      },
      "CS2 Server": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}